# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PJA-Sensei AI Microservice",
    "version": "0.1.0"
  },
  "paths": {
...
    "/conversations/{conversation_id}/messages": {
      "get": {
        "summary": "List Messages",
        "operationId": "list_messages_conversations__conversation_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Send Message",
        "operationId": "send_message_conversations__conversation_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/conversations/{conversation_id}/prelab": {
      "get": {
        "summary": "Get Prelab",
        "operationId": "get_prelab_conversations__conversation_id__prelab_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit Prelab",
        "operationId": "submit_prelab_conversations__conversation_id__prelab_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreLabSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}/prelab/generate": {
      "post": {
        "summary": "Generate Prelab",
        "operationId": "generate_prelab_conversations__conversation_id__prelab_generate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}/messages/stream": {
      "post": {
        "summary": "Send Message Stream",
        "operationId": "send_message_stream_conversations__conversation_id__messages_stream_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}/messages/{message_id}/regenerate": {
      "post": {
        "summary": "Regenerate Message",
        "operationId": "regenerate_message_conversations__conversation_id__messages__message_id__regenerate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/conversations/{conversation_id}/messages/{message_id}/feedback": {
      "post": {
        "summary": "Rate Message",
        "operationId": "rate_message_conversations__conversation_id__messages__message_id__feedback_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodeContext": {
        "properties": {
          "current_file_name": {
            "type": "string",
            "title": "Current File Name"
          },
          "current_code": {
            "type": "string",
            "title": "Current Code"
          },
          "error_logs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Logs"
          },
          "workspace_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Root"
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CodeSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticItem"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "open_files": {
            "items": {
              "$ref": "#/components/schemas/OpenFile"
            },
            "type": "array",
            "title": "Open Files"
          }
        },
        "type": "object",
        "required": [
          "current_file_name",
          "current_code"
        ],
        "title": "CodeContext"
      },
      "CodeSelection": {
        "properties": {
          "start_line": {
            "type": "integer",
            "title": "Start Line"
          },
          "end_line": {
            "type": "integer",
            "title": "End Line"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "start_line",
          "end_line"
        ],
        "title": "CodeSelection"
      },
      "DebugInfo": {
        "properties": {
          "is_frustrated": {
            "type": "boolean",
            "title": "Is Frustrated"
          },
          "avg_score": {
            "type": "number",
            "title": "Avg Score"
          },
          "code_changed": {
            "type": "boolean",
            "title": "Code Changed"
          }
        },
        "type": "object",
        "required": [
          "is_frustrated",
          "avg_score",
          "code_changed"
        ],
        "title": "DebugInfo"
      },
      "DiagnosticItem": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "severity": {
            "type": "string",
            "enum": [
              "error",
              "warning",
              "info",
              "hint"
            ],
            "title": "Severity",
            "default": "error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "DiagnosticItem"
      },
      "FeedbackRequest": {
        "properties": {
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "rating"
        ],
        "title": "FeedbackRequest"
      },
      "GoalProgressItem": {
        "properties": {
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "status": {
            "type": "string",
            "enum": [
              "not_started",
              "in_progress",
              "done"
            ],
            "title": "Status",
            "default": "not_started"
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "GoalProgressItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "code_context": {
            "$ref": "#/components/schemas/CodeContext"
          },
          "client_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Message Id"
          }
        },
        "type": "object",
        "required": [
          "question",
          "code_context"
        ],
        "title": "MessageRequest"
      },
      "MessageResponse": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "prompt_score": {
            "type": "integer",
            "title": "Prompt Score"
          },
          "prompt_feedback": {
            "type": "string",
            "title": "Prompt Feedback"
          },
          "tokens_used": {
            "type": "integer",
            "title": "Tokens Used"
          },
          "is_cached": {
            "type": "boolean",
            "title": "Is Cached",
            "default": false
          },
          "penalty_applied": {
            "type": "boolean",
            "title": "Penalty Applied",
            "default": false
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            },
            "type": "array",
            "title": "Sources"
          },
          "suggested_next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Next Step"
          },
          "goal_progress": {
            "items": {
              "$ref": "#/components/schemas/GoalProgressItem"
            },
            "type": "array",
            "title": "Goal Progress"
          },
          "debug_info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DebugInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Message Id"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "prompt_score",
          "prompt_feedback",
          "tokens_used"
        ],
        "title": "MessageResponse"
      },
      "OpenFile": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "path",
          "content"
        ],
        "title": "OpenFile"
      },
      "PreLabAnswerItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "id",
          "answer"
        ],
        "title": "PreLabAnswerItem"
      },
...
      "PreLabSubmitRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/PreLabAnswerItem"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "PreLabSubmitRequest"
      },
...
      "SourceRef": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url"
        ],
        "title": "SourceRef"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}